feat: Journa Client code optimization (#704)
* feat: Journa Client code optimization

---------
