+++ src/common/metadata-struct/src/adapter/record.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use common_base::tools::now_mills;
+use common_base::tools::now_second;
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
@@ -28,7 +28,7 @@ pub struct Record {
     pub key: String,
     pub data: Vec<u8>,
     pub tags: Vec<String>,
-    pub timestamp: u128,
+    pub timestamp: u64,
 }
 
 impl Record {
@@ -38,7 +38,7 @@ impl Record {
             key: "".to_string(),
             data,
             tags: Vec::new(),
-            timestamp: now_mills(),
+            timestamp: now_second(),
             header: Vec::new(),
         }
     }
@@ -49,7 +49,7 @@ impl Record {
             key: "".to_string(),
             data: data.as_bytes().to_vec(),
             tags: Vec::new(),
-            timestamp: now_mills(),
+            timestamp: now_second(),
             header: Vec::new(),
         }
     }
